site/requirements.txt
click==7.1.2
flask==1.1.2
itsdangerous==1.1.0
jinja2==2.11.2
markdown==3.2.2
markupsafe==1.1.1
pygments==2.6.1
pymdown-extensions==7.1
pyscss==1.3.7
pyyAML==5.3.1
six==1.15.0
werkzeug==1.0.1
